+++ lib/kuber_kit/core/configuration.rb
@@ -1,6 +1,7 @@
 class KuberKit::Core::Configuration
   attr_reader :name, :artifacts, :registries, :env_files, :templates, :kubeconfig_path, 
-              :services_attributes, :enabled_services, :disabled_services, :build_servers, :global_build_vars,
+              :services_attributes, :enabled_services, :disabled_services, :default_services, 
+              :build_servers, :global_build_vars,
               :deployer_strategy, :deployer_namespace, :deployer_require_confirimation
 
   Contract KeywordArgs[
@@ -13,6 +14,7 @@ class KuberKit::Core::Configuration
     services_attributes:  HashOf[Symbol => Hash],
     enabled_services:     ArrayOf[Symbol],
     disabled_services:    ArrayOf[Symbol],
+    default_services:     ArrayOf[Symbol],
     build_servers:        ArrayOf[KuberKit::Core::BuildServers::AbstractBuildServer],
     global_build_vars:    HashOf[Symbol => Any],
     deployer_strategy:              Symbol,
@@ -20,7 +22,8 @@ class KuberKit::Core::Configuration
     deployer_require_confirimation: Bool,
   ] => Any
   def initialize(name:, artifacts:, registries:, env_files:, templates:, kubeconfig_path:, 
-                 services_attributes:, enabled_services:, disabled_services:, build_servers:, global_build_vars:,
+                 services_attributes:, enabled_services:, disabled_services:, default_services:, 
+                 build_servers:, global_build_vars:,
                  deployer_strategy:, deployer_namespace:, deployer_require_confirimation:)
     @name                 = name
     @artifacts            = artifacts
@@ -32,6 +35,7 @@ class KuberKit::Core::Configuration
     @services_attributes  = services_attributes
     @enabled_services     = enabled_services
     @disabled_services    = disabled_services
+    @default_services     = default_services
     @global_build_vars    = global_build_vars
     @deployer_strategy              = deployer_strategy
     @deployer_namespace             = deployer_namespace