+++ spec/kuber_kit/core/configuration_definition_spec.rb
@@ -138,6 +138,15 @@ RSpec.describe KuberKit::Core::ConfigurationDefinition do
     end
   end
 
+  context "default_services" do
+    it "allows using array to set default services" do
+      subject.default_services([:service_1, :service_2])
+      subject.default_services([:service_3])
+
+      expect(subject.to_attrs.default_services).to eq([:service_1, :service_2, :service_3])
+    end
+  end
+
   context "service_attributes" do
     it "defines attributes for services" do
       subject.service_attributes(my_service: {scale: 1})